script: Split up `contenteditable.rs` into multiple files
This file was getting way too big and too cluttered. Instead,
split it up into multiple files in a dedicated folder.

Part of #25005
Part of #43709
